requirements.txt
mysql-connector-python-rf-2.2.2
Flask==0.10.1